# System IO refactor