--- a/src/Builder/BackgroundWriter.elm
+++ b/src/Builder/BackgroundWriter.elm
@@ -5,9 +5,9 @@ module Builder.BackgroundWriter exposing
     )
 
 import Builder.File as File
-import Data.IO as IO exposing (IO)
 import Json.Decode as Decode
 import Json.Encode as Encode
+import System.IO as IO exposing (IO)
 import Utils.Main as Utils