--- a/src/Builder/Build.elm
+++ b/src/Builder/Build.elm
@@ -214,7 +214,7 @@ type Module
 
 
 type alias Dependencies =
-    Dict ModuleName.Canonical I.DependencyInterface
+    Dict IO.Canonical I.DependencyInterface
 
 
 fromPaths : Reporting.Style -> FilePath -> Details.Details -> NE.Nonempty FilePath -> IO (Result Exit.BuildProblem Artifacts)
@@ -580,7 +580,7 @@ checkModule ((Env _ root projectType _ _ _ _) as env) foreigns resultsMVar name 
                         Error.BadSyntax err
 
         SForeign home ->
-            case Utils.find (ModuleName.Canonical home name) foreigns of
+            case Utils.find (IO.Canonical home name) foreigns of
                 I.Public iface ->
                     IO.pure (RForeign iface)
 
@@ -1270,7 +1270,7 @@ toDocs result =
 
 
 type ReplArtifacts
-    = ReplArtifacts ModuleName.Canonical (List Module) L.Localizer (Dict Name.Name Can.Annotation)
+    = ReplArtifacts IO.Canonical (List Module) L.Localizer (Dict Name.Name Can.Annotation)
 
 
 fromRepl : FilePath -> Details.Details -> String -> IO (Result Exit.Repl ReplArtifacts)
@@ -1356,7 +1356,7 @@ finalizeReplArtifacts ((Env _ root projectType _ _ _ _) as env) source ((Src.Mod
                         case result of
                             Ok (Compile.Artifacts ((Can.Module name _ _ _ _ _ _ _) as canonical) annotations objects) ->
                                 let
-                                    h : ModuleName.Canonical
+                                    h : IO.Canonical
                                     h =
                                         name