--- a/src/Builder/Elm/Details.elm
+++ b/src/Builder/Elm/Details.elm
@@ -50,6 +50,7 @@ import Data.Set as EverySet exposing (EverySet)
 import Json.Decode as Decode
 import Json.Encode as Encode
 import System.IO as IO exposing (IO)
+import System.TypeCheck.IO as TypeCheck
 import Utils.Crash exposing (crash)
 import Utils.Main as Utils exposing (FilePath, MVar)
 
@@ -101,7 +102,7 @@ type Extras
 
 
 type alias Interfaces =
-    Dict IO.Canonical I.DependencyInterface
+    Dict TypeCheck.Canonical I.DependencyInterface
 
 
 
@@ -447,7 +448,7 @@ addInterfaces directDeps pkg (Artifacts ifaces _) dependencyInterfaces =
     Dict.union ModuleName.compareCanonical
         dependencyInterfaces
         (Dict.fromList ModuleName.compareCanonical
-            (List.map (Tuple.mapFirst (IO.Canonical pkg))
+            (List.map (Tuple.mapFirst (TypeCheck.Canonical pkg))
                 (Dict.toList
                     (if Dict.member pkg directDeps then
                         ifaces