--- a/src/Builder/File.elm
+++ b/src/Builder/File.elm
@@ -14,11 +14,12 @@ module Builder.File exposing
     , zeroTime
     )
 
-import Data.IO as IO exposing (IO)
+import Codec.Archive.Zip as Zip
 import Json.Decode as Decode
 import Json.Encode as Encode
+import System.IO as IO exposing (IO(..))
 import Time
-import Utils.Main as Utils exposing (FilePath, ZipArchive, ZipEntry)
+import Utils.Main as Utils exposing (FilePath)
 
 
 
@@ -94,7 +95,7 @@ readBinary decoder path =
 
 writeUtf8 : FilePath -> String -> IO ()
 writeUtf8 path content =
-    IO.make (Decode.succeed ()) (IO.WriteString path content)
+    IO (\s -> ( s, IO.WriteString IO.pure path content ))
 
 
 
@@ -103,7 +104,7 @@ writeUtf8 path content =
 
 readUtf8 : FilePath -> IO String
 readUtf8 path =
-    IO.make Decode.string (IO.Read path)
+    IO (\s -> ( s, IO.Read IO.pure path ))
 
 
 
@@ -112,16 +113,16 @@ readUtf8 path =
 
 writeBuilder : FilePath -> String -> IO ()
 writeBuilder path builder =
-    IO.make (Decode.succeed ()) (IO.WriteString path builder)
+    IO (\s -> ( s, IO.WriteString IO.pure path builder ))
 
 
 
 -- WRITE PACKAGE
 
 
-writePackage : FilePath -> ZipArchive -> IO ()
+writePackage : FilePath -> Zip.Archive -> IO ()
 writePackage destination archive =
-    case Utils.zipZEntries archive of
+    case Zip.zEntries archive of
         [] ->
             IO.pure ()
 
@@ -129,17 +130,17 @@ writePackage destination archive =
             let
                 root : Int
                 root =
-                    String.length (Utils.zipERelativePath entry)
+                    String.length (Zip.eRelativePath entry)
             in
             Utils.mapM_ (writeEntry destination root) entries
 
 
-writeEntry : FilePath -> Int -> ZipEntry -> IO ()
+writeEntry : FilePath -> Int -> Zip.Entry -> IO ()
 writeEntry destination root entry =
     let
         path : String
         path =
-            String.dropLeft root (Utils.zipERelativePath entry)
+            String.dropLeft root (Zip.eRelativePath entry)
     in
     if
         String.startsWith "src/" path
@@ -151,7 +152,7 @@ writeEntry destination root entry =
             Utils.dirCreateDirectoryIfMissing True (Utils.fpForwardSlash destination path)
 
         else
-            writeUtf8 (Utils.fpForwardSlash destination path) (Utils.zipFromEntry entry)
+            writeUtf8 (Utils.fpForwardSlash destination path) (Zip.fromEntry entry)
 
     else
         IO.pure ()