--- a/src/Builder/Reporting/Exit/Help.elm
+++ b/src/Builder/Reporting/Exit/Help.elm
@@ -8,7 +8,6 @@ module Builder.Reporting.Exit.Help exposing
     , reportToJson
     , toStderr
     , toStdout
-    , toString
     )
 
 import Compiler.Json.Encode as E